Move xds tests to xds_test package
