+++ pilot/pkg/xds/ads.go
@@ -514,20 +514,6 @@ func warmingDependencies(typeURL string) []string {
 	}
 }
 
-// listEqualUnordered checks that two lists contain all the same elements
-func listEqualUnordered(a []string, b []string) bool {
-	if len(a) != len(b) {
-		return false
-	}
-	first := sets.New(a...)
-	for _, c := range b {
-		if !first.Contains(c) {
-			return false
-		}
-	}
-	return true
-}
-
 // update the node associated with the connection, after receiving a packet from envoy, also adds the connection
 // to the tracking map.
 func (s *DiscoveryServer) initConnection(node *core.Node, con *Connection, identities []string) error {
@@ -877,11 +863,11 @@ func (s *DiscoveryServer) AdsPushAll(req *model.PushRequest) {
 		}
 	}
 
-	s.startPush(req)
+	s.StartPush(req)
 }
 
 // Send a signal to all connections, with a push event.
-func (s *DiscoveryServer) startPush(req *model.PushRequest) {
+func (s *DiscoveryServer) StartPush(req *model.PushRequest) {
 	// Push config changes, iterating over connected envoys.
 	if log.DebugEnabled() {
 		currentlyPending := s.pushQueue.Pending()
