+++ pilot/pkg/xds/mesh_network_test.go
@@ -12,11 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package xds
+package xds_test
 
 import (
 	"context"
 	"fmt"
+	"istio.io/istio/pilot/pkg/xds"
 	"strings"
 	"testing"
 	"time"
@@ -42,6 +43,7 @@ import (
 	"istio.io/istio/pkg/config/schema/gvk"
 	"istio.io/istio/pkg/network"
 	"istio.io/istio/pkg/ptr"
+	"istio.io/istio/pkg/slices"
 	"istio.io/istio/pkg/test"
 	"istio.io/istio/pkg/test/util/retry"
 )
@@ -74,7 +76,7 @@ func TestNetworkGatewayUpdates(t *testing.T) {
 		configObjects = append(configObjects, w.configs()...)
 	}
 	meshNetworks := mesh.NewFixedNetworksWatcher(nil)
-	s := NewFakeDiscoveryServer(t, FakeOptions{
+	s := xds.NewFakeDiscoveryServer(t, xds.FakeOptions{
 		KubernetesObjects: kubeObjects,
 		Configs:           configObjects,
 		NetworksWatcher:   meshNetworks,
@@ -602,7 +604,7 @@ func runMeshNetworkingTest(t *testing.T, tt meshNetworkingTest, configs ...confi
 		}
 		configObjects = append(configObjects, w.configs()...)
 	}
-	s := NewFakeDiscoveryServer(t, FakeOptions{
+	s := xds.NewFakeDiscoveryServer(t, xds.FakeOptions{
 		KubernetesObjectsByCluster:      kubeObjects,
 		KubernetesObjectStringByCluster: tt.kubeObjectsYAML,
 		ConfigString:                    tt.configYAML,
@@ -660,7 +662,7 @@ func (w *workload) ExpectWithWeight(target *workload, subset string, eps ...xdst
 	w.weightedExpectations[target.clusterName(subset)] = eps
 }
 
-func (w *workload) Test(t *testing.T, s *FakeDiscoveryServer) {
+func (w *workload) Test(t *testing.T, s *xds.FakeDiscoveryServer) {
 	if w.expectations == nil && w.weightedExpectations == nil {
 		return
 	}
@@ -670,7 +672,7 @@ func (w *workload) Test(t *testing.T, s *FakeDiscoveryServer) {
 	})
 }
 
-func (w *workload) testUnweighted(t *testing.T, s *FakeDiscoveryServer) {
+func (w *workload) testUnweighted(t *testing.T, s *xds.FakeDiscoveryServer) {
 	if w.expectations == nil {
 		return
 	}
@@ -681,7 +683,7 @@ func (w *workload) testUnweighted(t *testing.T, s *FakeDiscoveryServer) {
 
 			for c, want := range w.expectations {
 				got := eps[c]
-				if !listEqualUnordered(got, want) {
+				if !slices.EqualUnordered(got, want) {
 					err := fmt.Errorf("cluster %s, expected %v, but got %v", c, want, got)
 					fmt.Println(err)
 					return err
@@ -689,7 +691,7 @@ func (w *workload) testUnweighted(t *testing.T, s *FakeDiscoveryServer) {
 			}
 			for c, got := range eps {
 				want := w.expectations[c]
-				if !listEqualUnordered(got, want) {
+				if !slices.EqualUnordered(got, want) {
 					err := fmt.Errorf("cluster %s, expected %v, but got %v", c, want, got)
 					fmt.Println(err)
 					return err
@@ -700,7 +702,7 @@ func (w *workload) testUnweighted(t *testing.T, s *FakeDiscoveryServer) {
 	})
 }
 
-func (w *workload) testWeighted(t *testing.T, s *FakeDiscoveryServer) {
+func (w *workload) testWeighted(t *testing.T, s *xds.FakeDiscoveryServer) {
 	if w.weightedExpectations == nil {
 		return
 	}
@@ -768,7 +770,7 @@ func (w *workload) configs() []config.Config {
 	return nil
 }
 
-func (w *workload) setupProxy(s *FakeDiscoveryServer) {
+func (w *workload) setupProxy(s *xds.FakeDiscoveryServer) {
 	p := &model.Proxy{
 		ID:     strings.Join([]string{w.name, w.namespace}, "."),
 		Labels: w.labels,
