+++ pilot/pkg/xds/sds.go
@@ -302,20 +302,6 @@ func filterAuthorizedResources(resources []SecretResource, proxy *model.Proxy, s
 	return allowedResources
 }
 
-func atMostNJoin(data []string, limit int) string {
-	if limit == 0 || limit == 1 {
-		// Assume limit >1, but make sure we dpn't crash if someone does pass those
-		return strings.Join(data, ", ")
-	}
-	if len(data) == 0 {
-		return ""
-	}
-	if len(data) < limit {
-		return strings.Join(data, ", ")
-	}
-	return strings.Join(data[:limit-1], ", ") + fmt.Sprintf(", and %d others", len(data)-limit+1)
-}
-
 func toEnvoyCaSecret(name string, certInfo *credscontroller.CertInfo) *discovery.Resource {
 	validationContext := &envoytls.CertificateValidationContext{
 		TrustedCa: &core.DataSource{
