+++ pilot/pkg/xds/util_test.go
@@ -1,0 +1,62 @@
+// Copyright Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package xds
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+)
+
+func TestAtMostNJoin(t *testing.T) {
+	tests := []struct {
+		data  []string
+		limit int
+		want  string
+	}{
+		{
+			[]string{"a", "b", "c"},
+			2,
+			"a, and 2 others",
+		},
+		{
+			[]string{"a", "b", "c"},
+			4,
+			"a, b, c",
+		},
+		{
+			[]string{"a", "b", "c"},
+			1,
+			"a, b, c",
+		},
+		{
+			[]string{"a", "b", "c"},
+			0,
+			"a, b, c",
+		},
+		{
+			[]string{},
+			3,
+			"",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(fmt.Sprintf("%s-%d", strings.Join(tt.data, "-"), tt.limit), func(t *testing.T) {
+			if got := atMostNJoin(tt.data, tt.limit); got != tt.want {
+				t.Errorf("got %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
