+++ pkg/slices/slices.go
@@ -32,6 +32,23 @@ func Equal[E comparable](s1, s2 []E) bool {
 	return slices.Equal(s1, s2)
 }
 
+// EqualUnordered reports whether two slices are equal, ignoring order
+func EqualUnordered[E comparable](s1, s2 []E) bool {
+	if len(s1) != len(s2) {
+		return false
+	}
+	first := make(map[E]struct{}, len(s1))
+	for _, c := range s1 {
+		first[c] = struct{}{}
+	}
+	for _, c := range s2 {
+		if _, f := first[c]; !f {
+			return false
+		}
+	}
+	return true
+}
+
 // EqualFunc reports whether two slices are equal using a comparison
 // function on each pair of elements. If the lengths are different,
 // EqualFunc returns false. Otherwise, the elements are compared in
